# api-docs.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getData","parameters":[],"responses":{"200":{"description":""}}}},"/api/users":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Users"]},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["Users"]}},"/api/users/{address}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"address","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Users"]},"patch":{"operationId":"UserController_update","parameters":[{"name":"address","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Users"]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"address","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["Users"]}},"/api/users/invite-code{address}":{"get":{"operationId":"UserController_generateInviteCode","parameters":[{"name":"address","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Users"]}},"/api/scores":{"post":{"operationId":"ScoreController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateScoreDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Score"}}}}},"tags":["Scores"]}},"/api/scores/{address}":{"get":{"operationId":"ScoreController_findAll","parameters":[{"name":"address","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Score"}}}}}},"tags":["Scores"]}},"/api/scores/{id}":{"get":{"operationId":"ScoreController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Score"}}}}},"tags":["Scores"]},"patch":{"operationId":"ScoreController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateScoreDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Score"}}}}},"tags":["Scores"]},"delete":{"operationId":"ScoreController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["Scores"]}},"/api/invite-code":{"post":{"operationId":"InviteCodeController_checkInviteCode","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckInviteCodeDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["InviteCode"]}}},"info":{"title":"Farty Bera API","description":"Farty Bera API Specification","version":"0.1","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{"address":{"type":"string"},"usedInviteCode":{"type":"string"}},"required":["address"]},"User":{"type":"object","properties":{"id":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string","nullable":true},"address":{"type":"string"},"inviteCode":{"type":"string"},"inviteCodeLimit":{"type":"number"},"usedInviteCode":{"type":"string"},"fartyHighScore":{"type":"number"},"fartyGamesPlayed":{"type":"number"},"honeyScore":{"type":"number"}},"required":["id","createdAt","updatedAt","deletedAt","address","inviteCode","fartyHighScore","fartyGamesPlayed","honeyScore"]},"UpdateUserDto":{"type":"object","properties":{"id":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string","nullable":true},"address":{"type":"string"},"inviteCode":{"type":"string"},"inviteCodeLimit":{"type":"number"},"usedInviteCode":{"type":"string"},"fartyHighScore":{"type":"number"},"fartyGamesPlayed":{"type":"number"},"honeyScore":{"type":"number"}}},"CreateScoreDto":{"type":"object","properties":{"value":{"type":"number"},"game":{"type":"object"},"address":{"type":"string"}},"required":["value","game","address"]},"Score":{"type":"object","properties":{"id":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string","nullable":true},"value":{"type":"number"},"game":{"type":"object"},"userAddress":{"type":"string"}},"required":["id","createdAt","updatedAt","deletedAt","value","game","userAddress"]},"UpdateScoreDto":{"type":"object","properties":{"value":{"type":"number"},"game":{"type":"object"},"address":{"type":"string"}}},"CheckInviteCodeDto":{"type":"object","properties":{"address":{"type":"string"},"inviteCode":{"type":"string"}},"required":["address","inviteCode"]}}}}