# apps/api/src/app/score/entities/score.entity.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsEnum, IsNumber, IsString } from 'class-validator';
import { Column, Entity } from 'typeorm';

import { Applications, BaseEntity } from '../../common';

@Entity('scores')
export class Score extends BaseEntity {
  @Column({ default: 0, type: 'int' })
  @ApiProperty()
  @IsNumber()
  value: number;

  @Column({ enum: Applications, type: 'enum' })
  @ApiProperty()
  @IsEnum(Applications)
  game: Applications;

  @Column({ type: 'text' })
  @ApiProperty()
  @IsString()
  userAddress: string;
}